+++ CHANGELOG.md
@@ -8,6 +8,12 @@
 
 ## v0.0
 
+### v0.0.0-5
+- 🍏 improve: Op ergonomics
+  - add type arg to `Op.loading<V>()` and `Op.error<V>(e)`
+  - add getters `op.isLoading`, `op.isReady`, `op.isError`
+  - add `error` getter, (also `pod.error`)
+
 ### v0.0.0-4
 - 🍏 fix: view types, view `.children` takes rest param for multiple children
 - 🍏 add: `use.life`